requirements.txt
openai
python-dotenv

#pip install -r requirements.txt

